+++ packages/block-editor/src/components/link-control/index.js
@@ -21,6 +21,7 @@ import LinkControlSearchInput from './search-input';
 import LinkPreview from './link-preview';
 import useCreatePage from './use-create-page';
 import { ViewerFill } from './viewer-slot';
+import { DEFAULT_LINK_SETTINGS } from './constants';
 
 /**
  * Default properties associated with a link control value.
@@ -104,8 +105,9 @@ import { ViewerFill } from './viewer-slot';
 function LinkControl( {
 	searchInputPlaceholder,
 	value,
-	settings,
+	settings = DEFAULT_LINK_SETTINGS,
 	onChange = noop,
+	onRemove,
 	noDirectEntry = false,
 	showSuggestions = true,
 	showInitialSuggestions,
@@ -189,6 +191,11 @@ function LinkControl( {
 		stopEditing();
 	};
 
+	const shownUnlinkControl =
+		onRemove && value && ! isEditingLink && ! isCreatingPage;
+
+	const showSettingsDrawer = !! settings?.length;
+
 	return (
 		<div
 			tabIndex={ -1 }
@@ -260,11 +267,25 @@ function LinkControl( {
 				/>
 			) }
 
-			<LinkControlSettingsDrawer
-				value={ value }
-				settings={ settings }
-				onChange={ onChange }
-			/>
+			{ ( showSettingsDrawer || shownUnlinkControl ) && (
+				<div className="block-editor-link-control__tools">
+					<LinkControlSettingsDrawer
+						value={ value }
+						settings={ settings }
+						onChange={ onChange }
+					/>
+					{ shownUnlinkControl && (
+						<Button
+							className="block-editor-link-control__unlink"
+							isDestructive
+							variant="link"
+							onClick={ onRemove }
+						>
+							{ __( 'Unlink' ) }
+						</Button>
+					) }
+				</div>
+			) }
 		</div>
 	);
 }