+++ packages/block-editor/src/components/link-control/link-preview.js
@@ -90,36 +90,43 @@ export default function LinkPreview( {
 				<ViewerSlot fillProps={ value } />
 			</div>
 
-			{ ( hasRichData || isFetching ) && (
+			{ ( ( hasRichData &&
+				( richData?.image || richData?.description ) ) ||
+				isFetching ) && (
 				<div className="block-editor-link-control__search-item-bottom">
-					<div
-						aria-hidden={ ! richData?.image }
-						className={ classnames(
-							'block-editor-link-control__search-item-image',
-							{
-								'is-placeholder': ! richData?.image,
-							}
-						) }
-					>
-						{ richData?.image && (
-							<img src={ richData?.image } alt="" />
-						) }
-					</div>
-					<div
-						aria-hidden={ ! richData?.description }
-						className={ classnames(
-							'block-editor-link-control__search-item-description',
-							{
-								'is-placeholder': ! richData?.description,
-							}
-						) }
-					>
-						{ richData?.description && (
-							<Text truncate numberOfLines="2">
-								{ richData.description }
-							</Text>
-						) }
-					</div>
+					{ ( richData?.image || isFetching ) && (
+						<div
+							aria-hidden={ ! richData?.image }
+							className={ classnames(
+								'block-editor-link-control__search-item-image',
+								{
+									'is-placeholder': ! richData?.image,
+								}
+							) }
+						>
+							{ richData?.image && (
+								<img src={ richData?.image } alt="" />
+							) }
+						</div>
+					) }
+
+					{ ( richData?.description || isFetching ) && (
+						<div
+							aria-hidden={ ! richData?.description }
+							className={ classnames(
+								'block-editor-link-control__search-item-description',
+								{
+									'is-placeholder': ! richData?.description,
+								}
+							) }
+						>
+							{ richData?.description && (
+								<Text truncate numberOfLines="2">
+									{ richData.description }
+								</Text>
+							) }
+						</div>
+					) }
 				</div>
 			) }
 		</div>