+++ packages/block-editor/src/components/rich-text/use-input-rules.js
@@ -60,7 +60,7 @@ export function useInputRules( props ) {
 		}
 
 		function onInput( event ) {
-			const { inputType } = event;
+			const { inputType, type } = event;
 			const {
 				value,
 				onChange,
@@ -69,7 +69,7 @@ export function useInputRules( props ) {
 			} = propsRef.current;
 
 			// Only run input rules when inserting text.
-			if ( inputType !== 'insertText' ) {
+			if ( inputType !== 'insertText' && type !== 'compositionend' ) {
 				return;
 			}
 
@@ -99,8 +99,10 @@ export function useInputRules( props ) {
 		}
 
 		element.addEventListener( 'input', onInput );
+		element.addEventListener( 'compositionend', onInput );
 		return () => {
 			element.removeEventListener( 'input', onInput );
+			element.removeEventListener( 'compositionend', onInput );
 		};
 	}, [] );
 }