+++ packages/block-library/src/embed/embed-preview.native.js
@@ -1,0 +1,83 @@
+/**
+ * External dependencies
+ */
+import { TouchableWithoutFeedback } from 'react-native';
+import { isEmpty } from 'lodash';
+
+/**
+ * WordPress dependencies
+ */
+import { View } from '@wordpress/primitives';
+
+import { BlockCaption } from '@wordpress/block-editor';
+import { __, sprintf } from '@wordpress/i18n';
+import { useState } from '@wordpress/element';
+
+/**
+ * Internal dependencies
+ */
+import EmbedNoPreview from './embed-no-preview';
+
+const EmbedPreview = ( {
+	clientId,
+	icon,
+	insertBlocksAfter,
+	isSelected,
+	label,
+	onBlur,
+	onFocus,
+} ) => {
+	const [ isCaptionSelected, setIsCaptionSelected ] = useState( false );
+
+	function accessibilityLabelCreator( caption ) {
+		return isEmpty( caption )
+			? /* translators: accessibility text. Empty Embed caption. */
+			  __( 'Embed caption. Empty' )
+			: sprintf(
+					/* translators: accessibility text. %s: Embed caption. */
+					__( 'Embed caption. %s' ),
+					caption
+			  );
+	}
+
+	function onEmbedPreviewPress() {
+		setIsCaptionSelected( false );
+	}
+
+	function onFocusCaption() {
+		if ( onFocus ) {
+			onFocus();
+		}
+		if ( ! isCaptionSelected ) {
+			setIsCaptionSelected( true );
+		}
+	}
+
+	return (
+		<TouchableWithoutFeedback
+			accessible={ ! isSelected }
+			onPress={ onEmbedPreviewPress }
+			disabled={ ! isSelected }
+		>
+			<View>
+				<EmbedNoPreview
+					label={ label }
+					icon={ icon }
+					isSelected={ isSelected }
+					onPress={ () => setIsCaptionSelected( false ) }
+				/>
+				<BlockCaption
+					accessibilityLabelCreator={ accessibilityLabelCreator }
+					accessible
+					clientId={ clientId }
+					insertBlocksAfter={ insertBlocksAfter }
+					isSelected={ isCaptionSelected }
+					onBlur={ onBlur }
+					onFocus={ onFocusCaption }
+				/>
+			</View>
+		</TouchableWithoutFeedback>
+	);
+};
+
+export default EmbedPreview;