+++ packages/components/src/dropdown-menu/index.native.js
@@ -89,7 +89,6 @@ function DropdownMenu( {
 				const openOnArrowDown = ( event ) => {
 					if ( ! isOpen && event.keyCode === DOWN ) {
 						event.preventDefault();
-						event.stopPropagation();
 						onToggle();
 					}
 				};