+++ packages/react-native-bridge/ios/GutenbergBridgeDelegate.swift
@@ -237,6 +237,8 @@ public protocol GutenbergBridgeDelegate: class {
     func gutenbergDidRequestMediaFilesUploadCancelDialog(_ mediaFiles: [[String: Any]])
 
     func gutenbergDidRequestMediaFilesSaveCancelDialog(_ mediaFiles: [[String: Any]])
+    
+    func gutenbergDidRequestPreview()
 }
 
 // MARK: - Optional GutenbergBridgeDelegate methods