+++ packages/react-native-bridge/ios/RNReactNativeGutenbergBridge.swift
@@ -356,6 +356,13 @@ public class RNReactNativeGutenbergBridge: RCTEventEmitter {
             self.delegate?.gutenbergDidSendButtonPressedAction(button)
         }
     }
+    
+    @objc
+    func requestPreview() {
+        DispatchQueue.main.async {
+            self.delegate?.gutenbergDidRequestPreview()
+        }
+    }
 
 }
 
@@ -382,6 +389,7 @@ extension RNReactNativeGutenbergBridge {
         case updateCapabilities
         case showNotice
         case mediaSave
+        case showEditorHelp
     }
 
     public override func supportedEvents() -> [String]! {